feat: fix comments and tests
