+++ fendermint/app/config/default.toml
@@ -15,6 +15,8 @@ builtin_actors_bundle = "bundle.car"
 custom_actors_bundle = "custom_actors_bundle.car"
 # Where to reach CometBFT for queries or broadcasting transactions.
 tendermint_rpc_url = "http://127.0.0.1:26657"
+# Where to reach CometBFT for subscriptions.
+tendermint_websocket_url = "ws://127.0.0.1:26657/websocket"
 # Block height where we should gracefully stop the node to perform maintenance or
 # with planning for an upcoming coordinated upgrade. Set to 0 to never halt.
 halt_height = 0
